+++ hax-lib/macros/src/implementation.rs
@@ -858,6 +858,17 @@ pub fn lean_pure_ensures_proof(payload: pm::TokenStream, item: pm::TokenStream) 
     quote! {#attr #item}.into()
 }
 
+/// This macro selects a proof method. This will influence the set of spec used for
+/// verification condition generation as well as the Lean tactic to discharge them.
+#[proc_macro_error]
+#[proc_macro_attribute]
+pub fn lean_proof_method(payload: pm::TokenStream, item: pm::TokenStream) -> pm::TokenStream {
+    let item: ItemFn = parse_macro_input!(item);
+    let payload = parse_macro_input!(payload as LitStr).value();
+    let attr = AttrPayload::ProofMethod(payload);
+    quote! {#attr #item}.into()
+}
+
 macro_rules! make_quoting_item_proc_macro {
     ($backend:ident, $macro_name:ident, $position:expr, $cfg_name:ident) => {
         #[doc = concat!("This macro inlines verbatim ", stringify!($backend)," code before a Rust item.")]