+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Changes to the Lean backend:
  - Extract correct `PhantomData` structure (#1932)
  - Standardize generated Lean naming to lowercase namespaces (#1914)
  - Fix associated constants with default values (#1941)
+ - New default proof for the Lean backend & proof method attribute (#1938)
 
 Miscellaneous: